requirements.txt
gunicorn>=20.1.0
jinja2>=3.1.2
PyJWT~=2.7.0